requirements.txt
django==3.2.6
psycopg2-binary==2.9.1
keras==2.9.0
tensorflow-macos==2.9.0
scikit-learn==2.9.0
pandas==1.5.3